pyproject.toml
[project]
name = "affective-memory-networks"
version = "1.0.0"
description = "Emotion-based memory for LLM agents - Official implementation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "numpy", "pandas", "matplotlib", "seaborn", "scikit-learn",
    "sentence-transformers", "bert-score", "scipy", "ollama", "tqdm"
]

[project.urls]
Homepage = "https://github.com/sakshyambanjade/Affective-Memory-Networks"
Repository = "https://github.com/sakshyambanjade/Affective-Memory-Networks"
